+++ app/scripts/lizard-nxt.js
@@ -427,5 +427,4 @@ angular.module('lizard-nxt')
     $('.navbar-version').toggle();
   };
 
-
 }]);